+++ week1/src/main/java/sopt/org/week1/advanced/Main.java
@@ -1,4 +1,4 @@
-package sopt.org.week1.bank;
+package sopt.org.week1.advanced;
 
 import java.util.ArrayList;
 import java.util.Scanner;
