+++ week1/src/main/java/sopt/org/week1/basic/Calculator.java
@@ -1,4 +1,4 @@
-package sopt.org.week1;
+package sopt.org.week1.basic;
 
 public class Calculator {
     private Character operator;
