+++ week1/src/main/java/sopt/org/week1/basic/Car.java
@@ -1,4 +1,4 @@
-package sopt.org.week1;
+package sopt.org.week1.basic;
 
 public interface Car {
     public abstract void turnOn();
