+++ week1/src/main/java/sopt/org/week1/basic/Dog.java
@@ -1,4 +1,4 @@
-package sopt.org.week1;
+package sopt.org.week1.basic;
 
 public class Dog extends Animal {
     private String gender;
