+++ week1/src/main/java/sopt/org/week1/basic/Person.java
@@ -1,4 +1,4 @@
-package sopt.org.week1;
+package sopt.org.week1.basic;
 
 // 추상화: 가장 본질적이고 공통적인 부분만 사용
 public abstract class Person {
