+++ week1/src/main/java/sopt/org/week1/basic/Student.java
@@ -1,4 +1,4 @@
-package sopt.org.week1;
+package sopt.org.week1.basic;
 
 public class Student extends Person {
